requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI
openai==1.3.0
httpx==0.27.0  # Pin httpx to avoid proxies parameter issue

# LangChain
langchain==0.1.20
langchain-openai==0.1.8
langchain-community==0.0.38

# RAG
faiss-cpu==1.7.4
numpy<2.0.0,>=1.19.3  # Pin numpy < 2.0 for compatibility with faiss-cpu and streamlit

# Text Processing
tiktoken==0.5.2  # For token-based chunking

# Data Processing
pandas==2.0.3

# Frontend
streamlit==1.28.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Document Processing
pdfplumber==0.10.0
beautifulsoup4==4.12.2

# Additional dependencies
python-multipart==0.0.6  # For file uploads in FastAPI

